+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Support python installation from Windows Store (#14)
+* Changed virtualenv version used by default from 16.7.9 to 20.4.2
+  (because only recent versions could work correctly with python installed from Windows Store)
+
 ### 2.2.0 (2020-04-06)
 * Add support for `pip` flags: [--trusted-host](https://pip.pypa.io/en/stable/reference/pip/#trusted-host) 
   and [--extra-index-url](https://pip.pypa.io/en/stable/reference/pip_install/#install-extra-index-url) (#10)